+++ include/nbla/function/randn.hpp
@@ -11,25 +11,25 @@ namespace Is
 {
     namespace nbla
     {
-        NBLA_REGISTER_FUNCTION_HEADER(Randn, float, float, const vector<int>&, int)
+        NBLA_REGISTER_FUNCTION_HEADER(Randn, float, float, const vector<int64_t>&, int)
         // float: Mean
         // float: Sigma
         // const vector<int>&: Shape
         // int: Seed
         
         template <typename T>
-        class Randn : public BaseFunction<float, float, const vector<int>&, int>
+        class Randn : public BaseFunction<float, float, const vector<int64_t>&, int>
         {
         protected:
             float mu_;
             float sigma_;
-            const vector<int> shape_;
+            const vector<int64_t> shape_;
             int seed_;
             std::mt19937 rgen_;
 
         public:
-            Randn(const Context& ctx, float mu, float sigma, const vector<int>& shape, int seed)
-                : BaseFunction<float, float, const vector<int>&, int>(ctx, mu, sigma, shape, seed)
+            Randn(const Context& ctx, float mu, float sigma, const vector<int64_t>& shape, int seed)
+                : BaseFunction<float, float, const vector<int64_t>&, int>(ctx, mu, sigma, shape, seed)
                 , mu_(mu)
                 , sigma_(sigma)
                 , shape_(shape)
